requirements.txt
Django==4.2
environs[django]==9.5.0
django-debug-toolbar==4.0.0
gunicorn==20.1.0
django-phonenumber-field==7.0.2
phonenumbers==8.13.9